[chore] move umap utils to a module
Allow the tests to be run from inside a cli, without requiring a browser.
