+++ package.json
@@ -40,8 +40,9 @@
     "@tmcw/togeojson": "^5.8.0",
     "colorbrewer": "^1.5.6",
     "csv2geojson": "5.1.1",
-    "dompurify": "^3.0.3",
+    "dompurify": "^3.0.11",
     "georsstogeojson": "^0.1.0",
+    "jsdom": "^24.0.0",
     "leaflet": "1.9.4",
     "leaflet-contextmenu": "^1.4.0",
     "leaflet-editable": "^1.2.0",
