+++ umap/static/umap/test/index.html
@@ -28,7 +28,6 @@
     <script src="../vendors/iconlayers/iconLayers.js" defer></script>
     <script src="../vendors/tokml/tokml.js" defer></script>
     <script src="../vendors/locatecontrol/L.Control.Locate.min.js" defer></script>
-    <script src="../vendors/dompurify/purify.min.js" defer></script>
     <script src="../vendors/colorbrewer/colorbrewer.js" defer></script>
     <script src="../vendors/simple-statistics/simple-statistics.min.js" defer></script>
