+++ umap/templates/umap/js.html
@@ -42,7 +42,6 @@
 <script src="{% static 'umap/vendors/tokml/tokml.js' %}" defer></script>
 <script src="{% static 'umap/vendors/locatecontrol/L.Control.Locate.min.js' %}"
         defer></script>
-<script src="{% static 'umap/vendors/dompurify/purify.min.js' %}" defer></script>
 <script src="{% static 'umap/vendors/colorbrewer/colorbrewer.js' %}" defer></script>
 <script src="{% static 'umap/vendors/simple-statistics/simple-statistics.min.js' %}"
         defer></script>
